Move EmptyState, Search, and Tree to separate files
